Improve coverage metrics
First, doing a trivial test around the "doStuff" command.
It does not, in fact, do anything, so the test is kinda dumb.

Second, want to exclude main.cpp in the test app from coverage
This required a bit of a shuffle (better design now) to include
the code coverage package at the top and then just call it throughout

Signed-off-by: Alex Shaver <alexshaver@gmail.com>